% Generated by build script
\name{ssb_tables}
\alias{ssb_tables}
\title{Browse SSB statistical tables}
\usage{
ssb_tables()
}
\arguments{
\item{path}{Path in the table tree. Empty string for root categories.}
}
\value{tibble: id, type ("l" = folder), text}
\description{Browse SSB statistical tables}
